Added 4 classes: block unit street postalcode to further split address
