+++ src/seedu/addressbook/data/person/Block.java
@@ -1,0 +1,15 @@
+package seedu.addressbook.data.person;
+
+public class Block {
+    private String block;
+
+    public Block (String block) {
+        this.block = block;
+    }
+    /**
+     * Returns String of block
+     */
+    public String getBlock() {
+        return block;
+    }
+}
