+++ src/seedu/addressbook/data/person/Street.java
@@ -1,0 +1,15 @@
+package seedu.addressbook.data.person;
+
+public class Street {
+    private String street;
+
+    public Street(String street) {
+        this.street = street;
+    }
+    /**
+     * Returns String of street
+     */
+    public String getStreet(){
+        return street;
+    }
+}
