+++ src/seedu/addressbook/data/person/Unit.java
@@ -1,0 +1,15 @@
+package seedu.addressbook.data.person;
+
+public class Unit {
+    private String unit;
+
+    public Unit(String unit) {
+        this.unit = unit;
+    }
+    /**
+     * Returns String of unit.
+     */
+    public String getUnit(){
+        return unit;
+    }
+}
